movie_recommender_api/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
pinecone-client
httpx
